refactor directory structure, need to clean up code
