+++ api/templates/api/blah.html
@@ -1,0 +1,1 @@
+Hello, world
