+++ osler/urls.py
@@ -14,5 +14,6 @@ urlpatterns = [
     url(r'^demographics/', include('demographics.urls')),
     url(r'^admin/', include(admin.site.urls)),
     url(r'^accounts/', include('django.contrib.auth.urls')),
+    url(r'^api/', include('api.urls')),
     url(r'^$', RedirectView.as_view(pattern_name="home", permanent=False)),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
