+++ pttrack/urls.py
@@ -36,27 +36,9 @@ unwrapped_urlpatterns = [  # pylint: disable=invalid-name
     url(r'^patient/activate_home/(?P<pk>[0-9]+)$',
         views.patient_activate_home,
         name='patient-activate-home'),
-    url(r'^pt_list/$',
-        views.PtList.as_view(),
-        name='pt_list_api'),
-    # url(r'^pt_list_last/$',
-    #     views.PtListLastName.as_view(),
-    #     name='pt_list_last_api'),
-    # url(r'^pt_list_latest/$',
-    #     views.PtListLatest.as_view(),
-    #     name='pt_list_latest_api'),
-    # url(r'^pt_list_active/$',
-    #     views.PtListActive.as_view(),
-    #     name='pt_list_active_api'),
-    # url(r'^pt_list_ai_active/$',
-    #     views.PtListActionItemListActive.as_view(),
-    #     name='pt_list_ai_active_api'),
-    # url(r'^pt_list_ai_inactive/$',
-    #     views.PtListActionItemListInactive.as_view(),
-    #     name='pt_list_ai_inactive_api'),
-    # url(r'^pt_list_unsigned/$',
-    #     views.PtListWorkupUnsigned.as_view(),
-    #     name='pt_list_unsigned_api'),
+    # url(r'^pt_list/$',
+    #     views.PtList.as_view(),
+    #     name='pt_list_api'),
 
     # PROVIDERS
     url(r'^new-provider/$',
